separate ftwalkcommon.c from dk functions
