+++ asm/melee/ft/chara/ftkirby.s
@@ -8975,7 +8975,7 @@ lbl_800F64C8:
 /* 800F64F8 000F30D8  C0 DF 01 20 */	lfs f6, 0x120(r31)
 /* 800F64FC 000F30DC  C0 FF 01 24 */	lfs f7, 0x124(r31)
 /* 800F6500 000F30E0  C1 07 00 7C */	lfs f8, 0x7c(r7)
-/* 800F6504 000F30E4  4B FE 97 A1 */	bl ftDonkey_800DFCA4
+/* 800F6504 000F30E4  4B FE 97 A1 */	bl ftWalkCommon_800DFCA4
 /* 800F6508 000F30E8  38 7F 00 00 */	addi r3, r31, 0
 /* 800F650C 000F30EC  38 80 01 FF */	li r4, 0x1ff
 /* 800F6510 000F30F0  4B F8 7D E5 */	bl func_8007E2F4
@@ -10715,7 +10715,7 @@ lbl_800F7E78:
 /* 800F7E78 000F4A58  7C 08 02 A6 */	mflr r0
 /* 800F7E7C 000F4A5C  90 01 00 04 */	stw r0, 4(r1)
 /* 800F7E80 000F4A60  94 21 FF F8 */	stwu r1, -8(r1)
-/* 800F7E84 000F4A64  4B FE 7F 59 */	bl ftDonkey_800DFDDC
+/* 800F7E84 000F4A64  4B FE 7F 59 */	bl ftWalkCommon_800DFDDC
 /* 800F7E88 000F4A68  80 01 00 0C */	lwz r0, 0xc(r1)
 /* 800F7E8C 000F4A6C  38 21 00 08 */	addi r1, r1, 8
 /* 800F7E90 000F4A70  7C 08 03 A6 */	mtlr r0
@@ -11099,7 +11099,7 @@ lbl_800F83D4:
 /* 800F83D4 000F4FB4  2C 00 00 00 */	cmpwi r0, 0
 /* 800F83D8 000F4FB8  40 82 00 58 */	bne lbl_800F8430
 /* 800F83DC 000F4FBC  7F C3 F3 78 */	mr r3, r30
-/* 800F83E0 000F4FC0  4B FE 78 91 */	bl ftDonkey_800DFC70
+/* 800F83E0 000F4FC0  4B FE 78 91 */	bl ftWalkCommon_800DFC70
 /* 800F83E4 000F4FC4  2C 03 00 00 */	cmpwi r3, 0
 /* 800F83E8 000F4FC8  41 82 00 48 */	beq lbl_800F8430
 /* 800F83EC 000F4FCC  83 BE 00 2C */	lwz r29, 0x2c(r30)
@@ -11115,7 +11115,7 @@ lbl_800F83D4:
 /* 800F8414 000F4FF4  C0 DD 01 20 */	lfs f6, 0x120(r29)
 /* 800F8418 000F4FF8  C0 FD 01 24 */	lfs f7, 0x124(r29)
 /* 800F841C 000F4FFC  C1 06 00 7C */	lfs f8, 0x7c(r6)
-/* 800F8420 000F5000  4B FE 78 85 */	bl ftDonkey_800DFCA4
+/* 800F8420 000F5000  4B FE 78 85 */	bl ftWalkCommon_800DFCA4
 /* 800F8424 000F5004  38 7D 00 00 */	addi r3, r29, 0
 /* 800F8428 000F5008  38 80 01 FF */	li r4, 0x1ff
 /* 800F842C 000F500C  4B F8 5E C9 */	bl func_8007E2F4
@@ -11414,7 +11414,7 @@ lbl_800F8848:
 /* 800F8848 000F5428  3C 60 80 0F */	lis r3, lbl_800F64C8@ha
 /* 800F884C 000F542C  38 83 64 C8 */	addi r4, r3, lbl_800F64C8@l
 /* 800F8850 000F5430  38 7E 00 00 */	addi r3, r30, 0
-/* 800F8854 000F5434  4B FE 76 75 */	bl ftDonkey_800DFEC8
+/* 800F8854 000F5434  4B FE 76 75 */	bl ftWalkCommon_800DFEC8
 lbl_800F8858:
 /* 800F8858 000F5438  80 01 00 2C */	lwz r0, 0x2c(r1)
 /* 800F885C 000F543C  83 E1 00 24 */	lwz r31, 0x24(r1)
@@ -11678,7 +11678,7 @@ lbl_800F8B90:
 /* 800F8B90 000F5770  7C 08 02 A6 */	mflr r0
 /* 800F8B94 000F5774  90 01 00 04 */	stw r0, 4(r1)
 /* 800F8B98 000F5778  94 21 FF F8 */	stwu r1, -8(r1)
-/* 800F8B9C 000F577C  4B FE 74 C5 */	bl ftDonkey_800E0060
+/* 800F8B9C 000F577C  4B FE 74 C5 */	bl ftWalkCommon_800E0060
 /* 800F8BA0 000F5780  80 01 00 0C */	lwz r0, 0xc(r1)
 /* 800F8BA4 000F5784  38 21 00 08 */	addi r1, r1, 8
 /* 800F8BA8 000F5788  7C 08 03 A6 */	mtlr r0
