+++ include/functions.h
@@ -2670,12 +2670,12 @@ void func_800DF0D0();
 // ? ftDonkey_800DFA70(?);
 // ? ftDonkey_800DFA98(?);
 // ? ftDonkey_800DFAE4(?);
-// ? ftDonkey_GetWalkType_800DFBF8(?);
-// ? ftDonkey_800DFC70(?);
-// ? ftDonkey_800DFCA4(?);
-// ? ftDonkey_800DFDDC(?);
-// ? ftDonkey_800DFEC8(?);
-// ? ftDonkey_800E0060(?);
+// ? ftWalkCommon_GetWalkType_800DFBF8(?);
+// ? ftWalkCommon_800DFC70(?);
+// ? ftWalkCommon_800DFCA4(?);
+// ? ftWalkCommon_800DFDDC(?);
+// ? ftWalkCommon_800DFEC8(?);
+// ? ftWalkCommon_800E0060(?);
 // ? ftDonkey_800E0134(?);
 // ? ftDonkey_800E017C(?);
 // ? ftDonkey_800E0294(?);
